Separated rules into `LocalVariableSnakeCase` and `ParameterSnakeCase`.
