+++ phpcs.xml
@@ -6,8 +6,9 @@
     <rule ref="Drupal"/>
     <!-- Checks that the strict_types has been declared. -->
     <rule ref="Generic.PHP.RequireStrictTypes" />
-    <!-- Apply our own DrevOps variable naming standard. -->
-    <rule ref="DrevOps.NamingConventions.VariableSnakeCase"/>
+    <!-- Apply our own DrevOps variable naming standards. -->
+    <rule ref="DrevOps.NamingConventions.LocalVariableSnakeCase"/>
+    <rule ref="DrevOps.NamingConventions.ParameterSnakeCase"/>
 
     <!-- Show sniff codes in all reports -->
     <arg value="s"/>
