+++ src/DrevOps/ruleset.xml
@@ -2,6 +2,9 @@
 <ruleset name="DrevOps">
     <description>DrevOps PHP_CodeSniffer standard.</description>
 
-    <!-- Enforce snake_case for local variables and function parameters -->
-    <rule ref="DrevOps.NamingConventions.VariableSnakeCase"/>
+    <!-- Enforce snake_case for local variables -->
+    <rule ref="DrevOps.NamingConventions.LocalVariableSnakeCase"/>
+
+    <!-- Enforce snake_case for function/method parameters -->
+    <rule ref="DrevOps.NamingConventions.ParameterSnakeCase"/>
 </ruleset>
